pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pos-local"
version = "1.0.0"
description = "POS minimal para Raspberry Pi (kiosk, ESC/POS, SQLite)"
requires-python = ">=3.9"
dependencies = [
  "PySide6",
  "python-escpos",
  "pyusb",
  "pydantic>=2",
  "typer",
  "rich",
  "argon2-cffi",
]

[project.scripts]
pos = "cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

